refactor: moved types to extra file
